Don't create duplicate fields.
When a field value on a subform is passed on the request, it would be rendered
as many times as the subform recurs on the layout. Or if a field  is present
in more than one subform, it would also be rendered multiple times. 

To prevent this, store seen field names on a volatile attribute on the form
being rendered.

For volatile attributes, see http://www.zope.org/Documentation/Articles/ZODB2

Perhaps there is a better way, but this is the best I could figure out. 


git-svn-id: https://plomino.svn.sourceforge.net/svnroot/plomino/trunk@1376 39f4f2d6-6c1e-0410-acfa-ed0e10f9290e