Add 'decimal' to PlominoUtils.
Note that methods to be called on these decimals need to be declared in PlominoUtils.
If a method returns a standard Decimal, it must be cast to a plomino decimal to call methods on it.
I.e. 'v.quantize(v).as_tuple()' will fail, but 'decimal(v.quantize(v)).as_tuple()' will succeed.