Refactor/clarify/simplify the user listing code.
The 'getFilteredNames' method returned [] if no filter.
This was changed to return the list unfiltered.
If you have no filter and want no users, you don't need to call the method.

Cleanups:
Factor out '_getNamesIds' used by 'getSelectionList' and 'getFilteredNames'.
The 'doc' parameter of 'getSelectionList' is not used. Give it a default
value so that you don't need to supply it.
Use meaningful names and tuple unpacking to make things clearer.
Sort early.
Log a degenerate case which probably indicates a problem in your Plomino app
(looking for nonexistent users).