Add a 'Hidden' widget for a text field.
Use this for fields that should not show up in the layout,
but that can compute a default value from the context and request.
Currently Plomino will inject hidden fields for request parameters that
match fields not in the layout, but this doesn't make it easy to grab
parameters that don't match fields, or combine parameters, or do some
computation to get the value.