A hidden text widget should be hidden.
If we render 'docid/OpenDocument?param=value', we may want to submit
the form to do some processing in a form event, even though we are not
saving the document. If we want 'param' to be passed along, we need to
render a hidden input for it even though we're not in edit mode.