Render the sums serverside, instead of letting datatables to this
If this approach is OK, we can remove fnFooterCallback from OpenView.
If someone wants to use it, they can add it in the field settings
(included via ${params}).

The generateTableFooter function could stay for use from such a
field settings usage.

This pattern would have to be documented with an example.